Change conditions and add a break a comments
Changed the condition for our while loop to run until there are no nodes left in the queue, and including a break when the node is at the destination location.
Moved around different actions like when the node is dequeued and we use a temp variable to point to it instead of leaving it in the queue and pointing at that node in the first index position of that array.
Changed the condition that builds our results output, and including an else to output a statement when a path cannot be found.